addendum to [1438915931bcdeda2350e1c85b1fa35d1bc0eff3] refs #21

+++ src/microsim/traffic_lights/MSTrafficLightLogic.cpp
@@ -157,7 +157,7 @@ MSTrafficLightLogic::init(NLDetectorBuilder&) {
                 }
                 if (!haveWarnedAboutUnusedStates && state1.size() > myLanes.size() + myIgnoredIndices.size()) {
                     WRITE_WARNINGF("Unused states in tlLogic '%', program '%' in phase % after tl-index %",
-                            getID(), getProgramID(), i, myLanes.size() - 1);
+                            getID(), getProgramID(), i, (int)myLanes.size() - 1);
                     haveWarnedAboutUnusedStates = true;
                 }
                 // detect illegal states
